+++ NEWS.md
@@ -1,6 +1,7 @@
 # pipapi (development version)
 ## New features
 - Add new SPR and SPL indicators
+- Add new endpoint `grouped-stats` to return grouped data
 
 ## Enhancements
 - Address some linting issues